move timerstate to own file
