+++ actix-http/Cargo.toml
@@ -92,6 +92,7 @@ criterion = { version = "0.3", features = ["html_reports"] }
 env_logger = "0.9"
 futures-util = { version = "0.3.7", default-features = false, features = ["alloc"] }
 memchr = "2.4"
+once_cell = "1.9"
 rcgen = "0.8"
 regex = "1.3"
 rustls-pemfile = "0.2"
