+++ actix-http/src/h1/dispatcher.rs
@@ -9,7 +9,7 @@ use std::{
 };
 
 use actix_codec::{AsyncRead, AsyncWrite, Decoder as _, Encoder as _, Framed, FramedParts};
-use actix_rt::time::{sleep_until, Instant, Sleep};
+use actix_rt::time::sleep_until;
 use actix_service::Service;
 use bitflags::bitflags;
 use bytes::{Buf, BytesMut};
@@ -28,6 +28,7 @@ use super::{
     codec::Codec,
     decoder::MAX_BUFFER_SIZE,
     payload::{Payload, PayloadSender, PayloadStatus},
+    timer::TimerState,
     Message, MessageType,
 };
 
@@ -162,98 +163,6 @@ pin_project! {
         read_buf: BytesMut,
         write_buf: BytesMut,
         codec: Codec,
-    }
-}
-
-#[derive(Debug)]
-enum TimerState {
-    Disabled,
-    Inactive,
-    Active { timer: Pin<Box<Sleep>> },
-}
-
-impl TimerState {
-    fn new(enabled: bool) -> Self {
-        if enabled {
-            Self::Inactive
-        } else {
-            Self::Disabled
-        }
-    }
-
-    fn is_enabled(&self) -> bool {
-        matches!(self, Self::Active { .. } | Self::Inactive)
-    }
-
-    fn set(&mut self, timer: Sleep, line: u32) {
-        if !self.is_enabled() {
-            log::warn!("setting disabled timer from line {}", line);
-        }
-
-        *self = Self::Active {
-            timer: Box::pin(timer),
-        };
-    }
-
-    #[allow(dead_code)]
-    fn reset(&mut self, deadline: Instant, line: u32) {
-        if !self.is_enabled() {
-            log::warn!("resetting disabled timer from line {}", line);
-        }
-
-        match self {
-            TimerState::Active { timer } => {
-                let _ = timer.as_mut().reset(deadline);
-            }
-
-            _ => self.set(sleep_until(deadline), line),
-        }
-    }
-
-    fn set_and_init(&mut self, cx: &mut Context<'_>, timer: Sleep, line: u32) {
-        self.set(timer, line);
-        self.init(cx);
-    }
-
-    fn clear(&mut self, line: u32) {
-        if !self.is_enabled() {
-            log::warn!("trying to clear a disabled timer from line {}", line);
-        }
-
-        if matches!(self, Self::Inactive) {
-            log::warn!("trying to clear an inactive timer from line {}", line);
-        }
-
-        *self = Self::Inactive;
-    }
-
-    fn init(&mut self, cx: &mut Context<'_>) {
-        if let TimerState::Active { timer } = self {
-            let _ = timer.as_mut().poll(cx);
-        }
-    }
-}
-
-impl fmt::Display for TimerState {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            TimerState::Disabled => f.write_str("timer is disabled"),
-            TimerState::Inactive => f.write_str("timer is inactive"),
-            TimerState::Active { timer } => {
-                let deadline = timer.deadline();
-                let now = Instant::now();
-
-                if deadline < now {
-                    f.write_str("timer is active and has reached deadline")
-                } else {
-                    write!(
-                        f,
-                        "timer is active and due to expire in {} milliseconds",
-                        ((deadline - now).as_secs_f32() * 1000.0)
-                    )
-                }
-            }
-        }
     }
 }
